Update README.md
Library App Assignment
Overview
Your task is to create a library app using Django. This app should allow users to store information
about books, authors, and book categories in a database, and track who has borrowed each book.
The app should also be able to assign a book to a borrower with a deadline to bring it back and
apply a fine if the borrower does not return the book on time. Finally, the app should allow users to
search for books and see if they are available.
Requirements
Your app should meet the following requirements:
1. Library staffs should be able to add, edit, and delete books from the database. Each book
should have a code, a title, a cover photo, and a list of authors and categories.
2. Library staffs should be able to register the borrowers and/or other staffs with their email
address.
3. Library staffs should be able to assign a book to a borrower with a deadline to bring it back.
If the book is not returned on time, a fine should be applied.
4. The app should have a page for library staff where they can see which books are available,
which books are currently borrowed, and which borrowers have overdue books.
5. The app should have a page for borrowers where they can log in to see which books they
have borrowed, when they are due, and any fines they may have incurred.
6. Users should be able to search for books by title, author, or category, and see if they are
available.
7. (Extra point) New users should get their random first password to their email address.
8. (Extra point) If the user would forgot his password there should be possibility to reset
password by getting verification to their email address.
Technical Details
Here are some technical details that should guide your implementation:
1. Use Django's built-in authentication system for user management and authentication.
2. Use Django's ORM to store and manage data in the database.
3. Use Django's class-based views to handle book creation, editing, and deletion, and
borrowing management.
4. Use Django's template system to render HTML pages.
5. Use Django's URL dispatcher to handle URL routing.
6. Use a CSS framework to make the app visually appealing.